Debugging MIDI Bluetooth
