+++ ShareplayMusicCollab/ShareplayMusicCollab/Model/MIDIEventType.swift
@@ -1,0 +1,16 @@
+//
+//  MIDIEventType.swift
+//  ShareplayMusicCollab
+//
+//  Created by Per Friis on 15/09/2024.
+//
+
+import Foundation
+
+enum MIDIEventType {
+    case none
+    case noteOn
+    case noteOff
+    case continuousControl
+    case programChange
+}
