+++ ShareplayMusicCollab/ShareplayMusicCollab/Model/MIDIMonitorData.swift
@@ -1,0 +1,21 @@
+//
+//  MIDIMonitorData.swift
+//  ShareplayMusicCollab
+//
+//  Created by Per Friis on 15/09/2024.
+//
+
+import Foundation
+
+struct MIDIMonitorData {
+    var noteOn = 0
+    var velocity = 0
+    var noteOff = 0
+    var channel = 0
+    var afterTouch = 0
+    var afterTouchNoteNumber = 0
+    var programChange = 0
+    var pitchWheelValue = 0
+    var controllerNumber = 0
+    var controllerValue = 0
+}
